+++ src/mqtt/index.js
@@ -1,3 +1,8 @@
+/*eslint strict: ["error", "global"]*/
+/*eslint no-invalid-this: "error"*/
+
+'use strict';
+
 module.exports = {
     mqtt: require('./mqtt_handlers.js'),
 };