+++ src/qrsUtil/index.js
@@ -1,3 +1,7 @@
+/*eslint strict: ["error", "global"]*/
+
+'use strict';
+
 module.exports = {
     senseStartTask: require('./senseStartTask.js'),
 };