+++ src/lib/post-to-influxdb.js
@@ -1,9 +1,14 @@
-const globals = require('../globals');
+/*eslint strict: ["error", "global"]*/
+/*eslint no-invalid-this: "error"*/
 
+'use strict';
+
+const globals = require('../globals');
+const  _ = require('lodash');
 
 
 function postButlerMemoryUsageToInfluxdb(memory) {
-    globals.logger.debug(`MEMORY USAGE: Memory usage ${JSON.stringify(memory, null, 2)})`);
+    // globals.logger.debug(`MEMORY USAGE: Memory usage ${JSON.stringify(memory, null, 2)})`);
 
     let datapoint = [
         {
@@ -14,13 +19,16 @@ function postButlerMemoryUsageToInfluxdb(memory) {
             fields: {
                 heap_used: memory.heapUsed,
                 heap_total: memory.heapTotal,
+                external: memory.external,
                 process_memory: memory.processMemory,
             },
         },
     ];
+    const deepClonedDatapoint = _.cloneDeep(datapoint);
 
     globals.influx
-        .writePoints(datapoint)
+        .writePoints(deepClonedDatapoint)
+
         .then(() => {
             globals.logger.silly(
                 `MEMORY USAGE: Influxdb datapoint for Butler memory usage: ${JSON.stringify(
@@ -30,6 +38,7 @@ function postButlerMemoryUsageToInfluxdb(memory) {
                 )}`,
             );
 
+            datapoint = null;
             globals.logger.verbose('MEMORY USAGE: Sent Butler memory usage data to InfluxDB');
         })
         .catch(err => {