+++ src/udp/index.js
@@ -1,3 +1,7 @@
+/*eslint strict: ["error", "global"]*/
+
+'use strict';
+
 module.exports = {
     udp: require('./udp_handlers.js'),
 };