+++ changelog.md
@@ -1,17 +1,35 @@
 # Change log
 
-## 4.1.0
+## 4.2.0
 
 ### New features
 
-- File copy method in the REST API.
-- Added Qlik Sense app "Butler 4.1 demo app" which contains examples of how Butler's API can be called from Sense load scripts.
+- Notifications by email when scheduled reload tasks fail. Notification email (both body and subject) are fully customizable using [Handlebars](https://handlebarsjs.com/guide/) templating syntax. [#92](https://github.com/ptarmiganlabs/butler/issues/92)
+- Notification emails include data such as task execution details, customizable number of rows from beginning and end of the reload script log, email priority, support for most email providers and more. [#92](https://github.com/ptarmiganlabs/butler/issues/92)
+- Email rate limits used to avoid spamming your inbox. [#92](https://github.com/ptarmiganlabs/butler/issues/92)
+- Notifications by email, Slack, MS Teams and MQTT when reload tasks are **aborted** in QMC. Email notifications use Handlebars templating syntax. New config file properties control which Slack/Teams channel these notifications are sent to. [#93](https://github.com/ptarmiganlabs/butler/issues/93)
+- Much improved documentation, especially in the [getting started/setup section of butler.ptarmiganlabs.com](https://butler.ptarmiganlabs.com/docs/getting-started/setup/).
+- Include non-heap memory in server uptime logging. Also store this metric to InfluxDB. [#100](https://github.com/ptarmiganlabs/butler/issues/100).
+
+### Fiexs and patches
+
+- Refactored code that starts Qlik Sense reload tasks. [#101](https://github.com/ptarmiganlabs/butler/issues/101)
+
+## 4.1.2
+
+### New features
 
 ### Fixes and patches
 
-- Updated dependencies to latest versions.
-- Minor stability changes/improvements to filemove and filedelete API methods.
-- Change parsing of config file to make Butler more forgiving if some parts of the config file are missing.
+- The included Sense demo app now correctly uses the Butler API endpoint used to start reload tasks.
+
+## 4.1.1
+
+### New features
+
+### Fixes and patches
+
+- API endpoint documentation for starting Sense reload tasks has been corrected.
 
 ### Breaking changes